+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changes
 Version 0.9.1 (UNRELEASED)
 --------------------------
 
-- Changes ``validate`` command to display non-critical validation warnings when checking the REANA specification file.
+- Adds support for Python 3.12.
 - Adds ``prune`` command to delete all intermediate files of a given workflow.
+- Changes ``validate`` command to display non-critical validation warnings when checking the REANA specification file.
 - Fixes ``list`` command to correctly list workflows when sorting them by their run number or by the size of their workspace.
 - Fixes ``du`` command help message typo.
 - Fixes ``validation --environments`` command to correctly handle fully qualified image names.